+++ app/src/main/java/de/passwordvault/view/activity_main/MainViewModel.java
@@ -1,10 +1,15 @@
 package de.passwordvault.view.activity_main;
 
+import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.lifecycle.ViewModel;
 import java.util.ArrayList;
 import de.passwordvault.model.entry.EntryAbbreviated;
 import de.passwordvault.model.entry.EntryManager;
+import de.passwordvault.model.rest.RestCallback;
+import de.passwordvault.model.rest.RestError;
+import de.passwordvault.model.rest.legal.LegalPageDto;
+import de.passwordvault.model.rest.legal.LegalRestClient;
 import de.passwordvault.model.storage.app.StorageException;
 
 
@@ -13,9 +18,20 @@ import de.passwordvault.model.storage.app.StorageException;
  * that shall be persistent throughout activity changes.
  *
  * @author  Christian-2003
- * @version 3.7.0
+ * @version 3.7.2
  */
 public class MainViewModel extends ViewModel {
+
+    /**
+     * Attribute stores the tag used with the REST client fetching the privacy policy.
+     */
+    public static final String TAG_PRIVACY = "privacy";
+
+    /**
+     * Attribute stores the tag used with the REST client fetching the terms of service.
+     */
+    public static final String TAG_TOS = "tos";
+
 
     /**
      * Attribute stores the adapter position of an entry that is currently opened from the MainActivity.
@@ -42,6 +58,35 @@ public class MainViewModel extends ViewModel {
      */
     private boolean loading;
 
+    /**
+     * Attribute stores the REST client used to fetch the privacy policy.
+     */
+    @NonNull
+    private final LegalRestClient privacyRestClient;
+
+    /**
+     * Attribute stores the REST client used to fetch the terms of service.
+     */
+    @NonNull
+    private final LegalRestClient tosRestClient;
+
+    /**
+     * Attribute stores the REST error occurred when fetching the privacy policy.
+     */
+    @Nullable
+    private RestError privacyError;
+
+    /**
+     * Attribute stores the REST error occurred when fetching the terms of service.
+     */
+    @Nullable
+    private RestError tosError;
+
+    /**
+     * Attribute indicates whether the legal page is displayed.
+     */
+    private boolean legalPageDisplayed;
+
 
     /**
      * Constructor instantiates a new MainViewModel with default values.
@@ -52,6 +97,11 @@ public class MainViewModel extends ViewModel {
         updateAvailable = false;
         loaded = false;
         loading = false;
+        privacyRestClient = new LegalRestClient(TAG_PRIVACY, "privacy");
+        tosRestClient = new LegalRestClient(TAG_TOS, "tos");
+        privacyError = null;
+        tosError = null;
+        legalPageDisplayed = false;
     }
 
 
@@ -105,9 +155,87 @@ public class MainViewModel extends ViewModel {
         updateAvailable = true;
     }
 
-
+    /**
+     * Method indicates whether data has changed since last cache generation.
+     *
+     * @return  Whether data has changed since last cache generation.
+     */
     public boolean isChangedSinceLastCacheGeneration() {
         return EntryManager.getInstance().isChangedSinceLastCacheGeneration();
+    }
+
+    /**
+     * Method returns the error which occurred while fetching the privacy policy.
+     *
+     * @return  REST error.
+     */
+    @Nullable
+    public RestError getPrivacyError() {
+        return privacyError;
+    }
+
+    /**
+     * Method changes the error which occurred while fetching the privacy policy.
+     *
+     * @param privacyError  REST error.
+     */
+    public void setPrivacyError(@Nullable RestError privacyError) {
+        this.privacyError = privacyError;
+    }
+
+    /**
+     * Method returns the error which occurred while fetching the terms of service.
+     *
+     * @return  REST error.
+     */
+    @Nullable
+    public RestError getTosError() {
+        return tosError;
+    }
+
+    /**
+     * Method changes the error which occurred while fetching the terms of service.
+     *
+     * @param tosError  REST error.
+     */
+    public void setTosError(@Nullable RestError tosError) {
+        this.tosError = tosError;
+    }
+
+    /**
+     * Method returns whether the legal page is displayed.
+     *
+     * @return  Whether the legal page is displayed.
+     */
+    public boolean isLegalPageDisplayed() {
+        return legalPageDisplayed;
+    }
+
+    /**
+     * Method changes whether the legal page is displayed.
+     *
+     * @param legalPageDisplayed    Whether the legal page is displayed.
+     */
+    public void setLegalPageDisplayed(boolean legalPageDisplayed) {
+        this.legalPageDisplayed = legalPageDisplayed;
+    }
+
+    /**
+     * Method returns the DTO to pass to {@link de.passwordvault.view.activity_legal.LegalActivity}.
+     *
+     * @return  DTO to pass to another activity.
+     */
+    public LegalPageDto getPrivacyDto() {
+        return privacyRestClient.toDto();
+    }
+
+    /**
+     * Method returns the DTO to pass to {@link de.passwordvault.view.activity_legal.LegalActivity}.
+     *
+     * @return  DTO to pass to another activity.
+     */
+    public LegalPageDto getTosDto() {
+        return tosRestClient.toDto();
     }
 
 
@@ -158,6 +286,17 @@ public class MainViewModel extends ViewModel {
             return null;
         }
         return EntryManager.getInstance().getData();
+    }
+
+
+    /**
+     * Method fetches the data from the REST API.
+     *
+     * @param callback  Callback to invoke once the data has been fetched.
+     */
+    public void fetchRestData(@Nullable RestCallback callback) {
+        privacyRestClient.fetch(callback);
+        tosRestClient.fetch(callback);
     }
 
 }